Cpp17 (#26)
* (cpp17) fixed compile issue when using C++17 standard.

* (cpp17) removing variant.h library as there is std::variant in cpp17

* (sonar) fixed vulnaribility warning with sprintf.

* (sonar) fixed overflow bug in Vertex::setName

* (sonar) fixed potential division by zeron in RoundRobinMapper

* (misc) set functions as static instead of const in CompilerExpression.

* (misc) renamed array_handle -> array_view. using std::unique_ptr for global stacks. fixed some sonar smells.

* (sonar) misc cleans.

* (fix) fixed doxygen url version.

* [cpp] fixed signed / unsigned warnings

* [cpp] fixed namespace with stackArray. fixed allocator-test.

* [cpp17] fixed Vertex::setName

* [cpp17] fixed seg fault in execDependenciesImpl.h. Fixed warnings in runtime-test

* [misc] fixing potential nullptr deferencing

* [misc] fixed GCC warning and usage of multithread during link time optim

---------

+++ test/runtime-test/appTest/reinforcement/include/clock.h
@@ -19,6 +19,6 @@
 void startTiming(int stamp);
 
 // Stoping to record time for a given stamp. Returns the time in us
-unsigned int stopTiming(int stamp);
+unsigned long stopTiming(int stamp);
 
 #endif
